+++ TP2-LaboPrograI/cValija.h
@@ -9,7 +9,7 @@ public:
 	/// Constructor parametrizado
 	/// </summary>
 	/// <param name="_peso">Recibira un peso (en kilogramos). El peso promedio de una valija es de 2.5 kilogramos </param>
-	cValija(float _peso = 2.5);
+	cValija(double _peso = 2.5);
 	/// <summary>
 	/// Destructor por defecto
 	/// </summary>
